+++ lib/ruby_spriter.rb
@@ -26,6 +26,8 @@ require_relative 'ruby_spriter/metadata_manager'
 require_relative 'ruby_spriter/video_processor'
 require_relative 'ruby_spriter/gimp_processor'
 require_relative 'ruby_spriter/consolidator'
+require_relative 'ruby_spriter/compression_manager'
+require_relative 'ruby_spriter/batch_processor'
 
 # Load orchestration
 require_relative 'ruby_spriter/processor'